tests/requirements.txt
allure-pytest==2.8.10
allure-python-commons==2.8.10
pytest==5.3.5
selenium==3.141.0

